% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhot.R
\name{sim.null.cross}
\alias{sim.null.cross}
\alias{include.hotspots}
\title{Simulate Null Dataset R/qtl Cross Object}
\usage{
sim.null.cross(
  chr.len = rep(400, 16),
  n.mar = 185,
  n.ind = 112,
  type = "bc",
  n.pheno = 6000,
  latent.eff = 1.5,
  res.var = 1,
  init.seed = 92387475
)

include.hotspots(
  cross,
  hotspotinfo,
  Q.eff,
  latent.eff,
  res.var = 1,
  n.pheno,
  init.seed
)
}
\arguments{
\item{chr.len}{vector of chromosome lengths}

\item{n.mar}{number of markers}

\item{n.ind}{number of individuals}

\item{type}{cross type}

\item{n.pheno}{}

\item{latent.eff}{latent effect}

\item{res.var}{}

\item{init.seed}{}

\item{cross}{R/qtl cross object}

\item{hotspotinfo}{data frame of hotspot info}

\item{Q.eff}{QTL effect}
}
\value{
R/qtl cross object

R/qtl cross object
}
\description{
Simulate Null Dataset R/qtl Cross Object

Generate hotspots for the simulated examples in the manuscript.
}
